+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 See this http://keepachangelog.com link for information on how we want this documented formatted.
 
+## v1.0.0
+
+#### Fixed
+
+* ALB query params & binary responses. Fixes #38.
+
+
 ## v0.6.0
 
 #### Added